\name{panel.pyramid}
\alias{panel.pyramid}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Default Panel Function for `pyramidlattice'
}
\description{
Default panel function for `pyramidlattice'. Essentially the same as `panel.barchart' in lattice package, but defaults to stacked bars.
}
\usage{
panel.pyramid(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{...}{see ?panel.barchart in lattice-package}
}
\seealso{
`barchart'
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
