\name{prepanel.default.bwplot2}
\alias{prepanel.default.bwplot2}
\title{
Default Prepanel Function for `pyramidlattice'
}
\description{
Default prepanel function for `pyramidlattice'. Essentially the same as `prepanel.default.bwplot' in lattice package, but defaults to stacked bars.
}
\usage{
prepanel.default.bwplot2( ...)
}
\arguments{
  \item{...}{for details refer to ?prepanel.default.bwplot in lattice-package}
}
\author{
Erich Striessnig <erich.striessnig@wu.ac.at>, modified from package lattice.
}
